+++ src/pandas_contract/mode.py
@@ -121,6 +121,10 @@ class Modes(enum.Enum):
             return self.value == other
         return super().__eq__(other)
 
+    def __hash__(self) -> int:
+        """Hash the mode."""
+        return hash(self.value)
+
 
 def get_mode() -> Modes:
     """Get the global mode for handling errors."""